+++ app/Models/Patient.php
@@ -8,7 +8,11 @@ use Illuminate\Database\Eloquent\Model;
 class Patient extends Model
 {
 
+
     use HasFactory;
+
+    protected $fillable=['name','email'];
+
     public function consultation(){
         return $this->hasMany(Consultation::class);
      }
