+++ resources/views/layouts/master.blade.php
@@ -91,6 +91,7 @@
 
           </div>
         </div>
+
       </div>
