+++ routes/web.php
@@ -18,8 +18,8 @@ use Illuminate\Support\Facades\Route;
 */
 
 
-// Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-// Auth::routes();
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+ Auth::routes();
 
 Route::get('/', function () {
     return view('welcome');
@@ -27,7 +27,7 @@ Route::get('/', function () {
 Route::get('/login/google', [App\Http\Controllers\Auth\LoginController::class, 'redirectToGoogle'])->name('login.google');
 Route::get('/login/google/callback', [App\Http\Controllers\Auth\LoginController::class, 'handleGoogleCallback']);
 
-Route::middleware(['auth', 'verified'])->group(function () {
+//Route::middleware(['auth', 'verified'])->group(function () {
     // ...
 
 
@@ -36,4 +36,5 @@ Route::get('planing/{id}',[RdvController::class,'index'])->name('rdv');
 Route::get('examen',[ExamenController::class,'index'])->name('examen');
 Route::get('suivi',[SuiviController::class,'index'])->name('suivi');
 Route::get('/dossier/patient/{id}',[DossierController::class,'pdf'])->name('dossier.patient');
-});
+Route::get('/rdv',[RdvController::class,'list'])->name('rdv.view');
+//});
